requirements.txt
Django==4.0
Pillow==9.0.1
django-taggit==2.1.0